fix bad local docker group detection
On prod deployment, there are actually 2 groups starting with 'docker'
so the old detection actually got '994\n995'.

$ grep docker /etc/group
dockerroot:x:995:
docker:x:994:

This result in the group not created properly inside the jenkins slave
container:

$ docker logs jenkins-config-prod_slave_1
+ [ -n 995
994 ]
+ addgroup --gid 995 994 docker_from_host
addgroup: The user `994' does not exist.
+ usermod -G docker_from_host -a jenkins
usermod: group 'docker_from_host' does not exist

The impact is the jenkins slave is unable to use docker during the build
since the user jenkins inside the container was not part of the same
docker group as the docker host.  Unable to perform any build.