add title slide, mention using notebooks as tests and tutorials, automatic slide numbering
Also worth mentionning that there is no way to save and use the slides
in complete offline mode.

I've tried many Firefox plugins (SingleFile, Save Page WE, WebScrapBook)
and none of them are able to save locally the multiple javascripts
properly so that we can launch the slides without internet connectivity.

Since we need internet connectivity anyways, I've changed the path of
the animated gif images to online as well.

At the presentation, we will need to preload the browser from home to
not rely on internet connection at the presentation kiosk.

Reveal.js does support automatic slide numbering but the configuration
setting is not exposed when we convert Jupyter notebook to slide so I
had to write a script to alter the converted .html file.